+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change KeePassXC to be installed as a Flatpak.
 - Change LibreOffice to be installed as a Flatpak.
 - Change PeaZip to be installed as a Flatpak.
+- Change OBS Studio to be installed as a Flatpak.
 
 ## 3.0.1